requirements.txt
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
dateparser==0.7.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.0.0
urllib3==1.25.9
zope.interface==5.1.0
